fix: move setuptools integration tests out of vcs-versioning
Move tests that require setuptools_scm (use_scm_version subprocess tests)
from vcs-versioning/testing_vcs/test_git.py to setuptools-scm/testing_scm/.

Also fix the NonNormalizedVersion string reference to use
vcs_versioning.NonNormalizedVersion instead of setuptools_scm.NonNormalizedVersion,
avoiding a runtime import of setuptools_scm.

Part of #1353
